server/requirements.txt
# === Core Dependencies (required) ===
setuptools>=69.0.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pillow==10.4.0
numpy==1.26.4
pydub==0.25.1
aiofiles==24.1.0
pyyaml>=6.0
aiosqlite>=0.19.0
apscheduler>=3.10.0

# === LLM / Agent ===
langchain==0.3.0
langchain-community==0.3.0

# === Google Services (optional but lightweight) ===
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.147.0
imapclient==3.0.1
